refactor: パッケージ構成修正
refs #002
