+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/web/controller/mobilesuit/MobileSuitQuery.java
@@ -2,7 +2,6 @@ package jp.co.sunarch.mobilesuitDatabase.port.adapter.web.controller.mobilesuit;
 
 import java.util.List;
 
-import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.from.mobilesuit.MobileSuitDetailFrom;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.mobilesuit.MobileSuitDetailForm;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.mobilesuit.MobileSuitForm;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.result.mobilesuit.EditMobileSuitResult;
@@ -11,7 +10,7 @@ public interface MobileSuitQuery {
 	
 	public List<MobileSuitForm> getMobileSuitList();
 
-	public MobileSuitDetailForm getMobileSuitDetail(MobileSuitDetailFrom msIdFrom);
+	public MobileSuitDetailForm getMobileSuitDetail(String msId);
 
 	public EditMobileSuitResult getMobileSuitId(String msId);
